moved the generate_sph_field method to the arbitrary grid class.
